major improvement to verification
This drastically reduces the amount of duplicate code and simplifies the
process of verification. This also makes it much easier to add
verification before critical actions (like an email change).

At a high level, this makes a route for verifications in general that
anything can send the user to. It also provides handy reusable utilities
for the purpose. This also adds a resource route for doing inline
verifications as well.